+++ Sources/Scaletor/Extensions/Accidental+CustomStringConvertible.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+extension Accidental: CustomStringConvertible {
+    public var description: String {
+        switch self {
+        case .doubleFlat:
+            return "♭♭"
+        case .flat:
+            return "♭"
+        case .natural:
+            return ""
+        case .sharp:
+            return "♯"
+        case .doubleSharp:
+            return "♯♯"
+        }
+    }
+}